+++ forui/lib/src/localizations/localizations_sq.dart
@@ -141,7 +141,7 @@ class FLocalizationsSq extends FLocalizations {
   String get timeFieldTimeSeparator => ':';
 
   @override
-  String get timeFieldPeriodSeparator => ' ';
+  String get timeFieldPeriodSeparator => ' ';
 
   @override
   String get timeFieldSuffix => '';