+++ forui/lib/src/widgets/time_field/input/time_input_controller.dart
@@ -13,6 +13,7 @@ import 'package:forui/src/widgets/time_field/input/time_parser.dart';
 abstract class TimeInputController extends InputController {
   final FTimeFieldController controller;
   final DateFormat format;
+  final String _canonicalSpace;
 
   factory TimeInputController(
     FLocalizations localizations,
@@ -51,8 +52,16 @@ abstract class TimeInputController extends InputController {
     super.style,
     super.parser,
     super.placeholder,
-  ) {
+  ): _canonicalSpace = format.canonicalSpace {
     controller.addListener(updateFromTimeController);
+  }
+
+  @override
+  set value(TextEditingValue newValue) {
+    if (_canonicalSpace != ' ' && newValue.text.contains(' ')) {
+      newValue = newValue.copyWith(text: newValue.text.replaceAll(' ', _canonicalSpace));
+    }
+    super.value = newValue;
   }
 
   // Used for input-driven changes only.