+++ forui/test/src/widgets/date_field/input/input_date_field_test.dart
@@ -17,7 +17,7 @@ void main() {
       (null, 'MM/DD/YYYY'), // M/d/y
       (Locale('en', 'SG'), 'DD/MM/YYYY'), // dd/MM/y
       (Locale('hr'), 'DD. MM. YYYY.'),
-      (Locale('bg'), 'DD.MM.YYYY г.'),
+      (Locale('bg'), 'DD.MM.YYYY г.'),
     ].indexed) {
       testWidgets('placeholder - $description - $index', (tester) async {
         await tester.pumpWidget(TestScaffold.app(locale: locale, child: field));